# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/get/{id}": {
            "get": {
                "description": "Fetches a job by its ID with partial content handling",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Retrieve a job from the queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved job",
                        "schema": {
                            "$ref": "#/definitions/jobs.Job"
                        }
                    },
                    "206": {
                        "description": "Partially completed job",
                        "schema": {
                            "$ref": "#/definitions/jobs.Job"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/errors.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.RestErr"
                        }
                    }
                }
            }
        },
        "/api/upload": {
            "post": {
                "description": "Validates and creates a new job in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Upload a new job to the queue",
                "parameters": [
                    {
                        "description": "Job to be uploaded",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jobs.Job"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Job successfully created",
                        "schema": {
                            "$ref": "#/definitions/jobs.Job"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "$ref": "#/definitions/errors.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.RestErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.RestErr": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "jobs.Job": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                },
                "Input": {
                    "type": "string"
                },
                "Slurml": {
                    "type": "boolean"
                },
                "lastUpdated": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "slurmID": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}